pyproject.toml
[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"

[project]
name = "deal-flow-engine"
version = "0.1.0"
description = "Private deal flow intelligence engine"
requires-python = ">=3.11"

# Runtime dependencies — only what Slice 1 needs today.
# Each later slice will add to this list with its own justification.
dependencies = [
    # psycopg2-binary: Postgres driver for direct SQL execution (no ORM).
    # Binary distribution avoids requiring libpq-dev at install time.
    "psycopg2-binary>=2.9",

    # google-api-python-client: Official Google API client. Required by Slice 4
    # (Gmail ingestion) and Slice 5 (Calendar ingestion) for API discovery and calls.
    "google-api-python-client>=2.100",

    # google-auth: Core Google auth library. Used directly for Credentials and
    # token refresh in gmail_connector.py. Transitive dep of google-auth-oauthlib
    # but imported directly, so declared explicitly.
    "google-auth>=2.0",

    # google-auth-oauthlib: OAuth2 credential handling for Google APIs.
    # Loads pre-existing token.json with refresh token for headless operation.
    "google-auth-oauthlib>=1.2",

    # python-dotenv: Loads .env files into os.environ so DATABASE_URL,
    # GOOGLE_CREDENTIALS_PATH, etc. don't need to be exported manually.
    "python-dotenv>=1.0",

    # anthropic: Official Anthropic Python SDK. Used by Slice 6 (entity extraction)
    # to call the Messages API for structured data extraction from emails/events.
    "anthropic>=0.40",

    # jinja2: Template engine for rendering the daily brief HTML before PDF conversion.
    # Used by Slice 9 (PDF renderer) to produce structured recommendation cards.
    "jinja2>=3.1",

    # weasyprint: HTML-to-PDF renderer with CSS print support. No Chromium dependency.
    # Used by Slice 9 to produce the daily brief PDF optimized for grayscale printing.
    "weasyprint>=62.0",
]

[project.optional-dependencies]
dev = [
    # pytest: test runner; required by `make test`.
    "pytest>=8.0",

    # ruff: fast Python linter and formatter; replaces flake8 + isort + pyupgrade
    # in a single zero-config binary. Chosen over flake8 for speed and unified config.
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = []
